requirements.txt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.1
idna==2.8
nltk==3.4
numpy==1.16.2
praw==6.1.1
prawcore==1.0.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.18
textblob==0.15.3
update-checker==0.16
urllib3==1.24.1
websocket-client==0.55.0